Add files via upload
Fixed a bug caused by be slicing the data hard by 300, instead of using all data and just slice for a subset prediction.
Also I missed init models with static random_seed.
As it seem it didnt actually impact the results that much, but affected the review of every dependant evaluation. 
Also elaborating more about what a forest with trees is and why random. :)